fixed 729b835 from https://gitee.com/life-liu/telephony_ril_adapter/pulls/235 IssueNo: codex修改
Description: codex修改
Sig: Sol_Enhanced_Telephony
Feature or Bugfix: Bugfix
Binary Source: No

+++ frameworks/src/hril_data_parcel.cpp
@@ -134,7 +134,7 @@ bool DataCallResultList::ReadFromParcel(Parcel &parcel)
         return false;
     }
     dcList.resize(size);
-    for (int32_t i = 0; i < size; i++) {
+    for (int32_t i = 0; i < (int32_t)size; i++) {
         dcList[i].ReadFromParcel(parcel);
     }
     return true;
